:truck: Move some rental car components to correct location.
Component folders should match model names where possible.
